+++ src/Plainion.GraphViz/IGraphViewExport.cs
@@ -1,0 +1,9 @@
+﻿using System.IO;
+
+namespace Plainion.GraphViz
+{
+    public interface IGraphViewExport
+    {
+        void ExportAsPng(Stream stream);
+    }
+}